+++ mesos/server.go
@@ -243,6 +243,8 @@ func IsK3SServerRunning() bool {
 		return false
 	}
 
+	defer res.Body.Close()
+
 	if res.StatusCode != 200 {
 		logrus.Error("IsK3SServerRunning: Error Status is not 200")
 		return false