+++ mesos/agent.go
@@ -13,33 +13,43 @@ import (
 )
 
 // SearchMissingK3SAgent Check if all agents are running. If one is missing, restart it.
-func SearchMissingK3SAgent() {
+func SearchMissingK3SAgent(restart bool) bool {
+	status := make([]mesosproto.TaskState, config.K3SAgentMax)
+	allRunning := true
 	if config.State != nil {
 		for i := 0; i < config.K3SAgentMax; i++ {
 			state := StatusK3SAgent(i)
 			if state != nil {
+				status[i] = *state.Status.State
 				if *state.Status.State != mesosproto.TASK_RUNNING {
+					allRunning = false
 					logrus.Debug("Missing K3SAgent: ", i)
-					StartK3SAgent(i)
+					if restart {
+						StartK3SAgent(i)
+					}
 				}
+			} else {
+				allRunning = false
 			}
 		}
+	} else {
+		allRunning = false
 	}
+	config.M3SStatus.Agent = status
+	return allRunning
 }
 
-// StatusK3SAgent Get out Status of the given agent ID
+// StatusK3SAgent Get out Status of the given agent
 func StatusK3SAgent(id int) *cfg.State {
 	if config.State != nil {
 		for _, element := range config.State {
 			if element.Status != nil {
-				if element.Command.InternalID == id && element.Command.IsK3SAgent == true {
-					config.M3SStatus.Agent = element.Status.State
+				if element.Command.InternalID == id && element.Command.IsK3SAgent {
 					return &element
 				}
 			}
 		}
 	}
-	config.M3SStatus.Agent = mesosproto.TASK_UNKNOWN.Enum()
 	return nil
 }