+++ changelog.md
@@ -5,6 +5,7 @@
 - Add status information about the M3s services.
 - Mesos-CLI: Get out the M3s status information.
 - Add Mesos Label support to (as example) control traefik
+- Add possibility to run the K8 Server and Agent on a specified hostname
 
 ## v0.1.0