+++ changelog.md
@@ -12,6 +12,7 @@
 - Add FrameworkName to Port Label
 - Add Kubernetes Status to Framework, bootstrap Server and Mesos-CLI
 - Add the possibility to set custom environment variables for the bootstrap.sh
+- Remove unneeded API calls
 
 ## v0.1.0